+++ packages/components/elements/command/Command.tsx
@@ -175,7 +175,7 @@ const CommandShortcut = ({
   );
 };
 
-type FullCommandItem = {
+export type FullCommandItem = {
   type: "group" | "separator";
   heading?: string;
   items?: {
@@ -239,7 +239,8 @@ const AppCommand = ({ commandProps, dialogProps }: AppCommandProps) => {
     <Dialog {...dialogProps}>
       <DialogContent
         hideCloseButton
-        className="hawa-overflow-hidden !hawa-p-0 hawa-shadow-lg !hawa-min-h-[50%]"
+        className="hawa-overflow-hidden !hawa-p-0 hawa-shadow-l "
+        // !hawa-min-h-[50%]
       >
         <FullCommand {...commandProps} />
       </DialogContent>